requirements.txt
# LangChain stack
langchain-core==0.3.66
langchain==0.3.26
langchain-community==0.3.26
langchain-ibm==0.3.13
ibm-watsonx-ai==1.3.26

# Core dependency
numpy>=1.26.0

# UI Layer
streamlit==1.36.0

# Environment variables
python-dotenv==1.0.1

